feat(core-executor): widen clk to a global, never-reset u64
`clk` was shard-local: reset to 0 at every shard cut and range-checked
to 24 bits, capping a shard at CORE_SHARD_CLK_LIMIT cycles. It's now a
global u64 that never resets across shards, matching upstream sp1's
bf3c7c6b5/a8e427c28 transition (see the AIR-side redesign in the
following commit for the clk_high/clk_low split this enables).

- ExecutionState::clk: u32 -> u64, plus a new initial_timestamp
  tracking the current shard's starting clk (used both for the
  cycles-this-shard shard-cut check and to replace the old
  record.shard != shard local-memory-event detection).
- Removed MemoryRecord/MemoryReadRecord/MemoryWriteRecord/
  MemoryInitializeFinalizeEvent's shard fields and every InstrEvent-
  family struct's shard field -- redundant now that clk alone
  identifies ordering.
- Shard-cutting decoupled from clk entirely: the old
  CORE_SHARD_CLK_LIMIT-based cut is gone, since clk no longer bounds
  a shard's size.
- New BumpClkHighEvent, emitted whenever an instruction's clk
  advance would leave insufficient room in the current 1<<24 window
  for its own sub-position offsets (MemoryAccessPosition, up to 4) --
  the executor pre-emptively jumps clk to the next window boundary
  first. Consumed by the new StateBumpChip AIR in the next commit.

+++ crates/core/executor/src/events/cpu.rs
@@ -14,7 +14,7 @@ use super::memory::MemoryRecordEnum;
 #[derive(Debug, Copy, Clone, Serialize, Deserialize)]
 pub struct CpuEvent {
     /// The clock cycle.
-    pub clk: u32,
+    pub clk: u64,
     /// The program counter.
     pub pc: u32,
     /// The next program counter.
@@ -49,7 +49,7 @@ pub struct CpuEvent {
 #[repr(C)]
 pub struct CpuEventFfi {
     /// The clock cycle.
-    pub clk: u32,
+    pub clk: u64,
     /// The program counter.
     pub pc: u32,
     /// The next program counter.
